requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1

sentence-transformers==2.2.2
transformers>=4.36.0,<4.42.0
torch==2.0.1

llama-index==0.10.0
llama-index-core==0.10.0
llama-index-vector-stores-supabase==0.1.0
llama-index-embeddings-huggingface==0.1.0
huggingface-hub==0.23.4

psycopg2-binary==2.9.9

PyPDF2==3.0.1
pypdf==3.17.0
python-docx==1.1.0

python-dotenv==1.0.0
requests==2.31.0
numpy>=1.25.0,<2.0
pandas==2.1.3
sqlalchemy==2.0.23

gunicorn==21.2.0